Implement StereoScreenShot to replace SimpleScreenShot
This uses the reverse stereo blit control in nvapi to get access to the
second back buffer and create a stereo screenshot.

Documentation on the reverse stereo blit is here - this does not list
which method to use for the blit in DX11, but there was really only one
possibility (CopySubresourceRegion), which worked fine:

http://docs.nvidia.com/gameworks/content/gameworkslibrary/coresdk/nvapi/group__stereoapi.html#gabf25d1ff8c91bc13fe42f097e0e5ae05

SimpleScreenShot is left in the code, but currently disabled. Both of
these methods together provide good examples of the process to get a
screenshot, and it might make sense to switch between the two based on
whether the source Texture2D is stereo or not, but short of checking if
stereo is enabled I'm not sure how to check that.