Remove throttle in RunFrameActions()
This throttle was originally added to reduce the number of times per
frame that we would attempt to process input (which was identified as a
performance bottleneck), however now that RunFrameActions is called from
Present(), we know that it will only be called once per frame so this
throttle is no longer required.

Remove it to be sure that RunFrameActions() will run every frame so that
anything we call from it can depend on that. Will be shortly become
important for the upcoming frame analysis support.