Save .JPS in addition to .DDS for each render target during frame analysis
The DDS plugin for the GIMP is failing to open many of the render
targets with formats that it doesn't support. For now the easiest path
forward is to dump out JPS files for each render target as well.

Not all texture formats get exported (in particular, depth targets fail
to export as JPS), but that's fine since we have the DDS files for the
rest.