Add keys to kill various DX11 features with only limited 3Dmigoto support
Adds keys that kill these DX11 features while held down:

- Compute shaders
- Geometry shaders
- Tessellation (hull & domain shaders)
- Deferred contexts / command lists

This allows us to quickly check if any of these features are in use, and
what they are responsible for. The disabling may not be thorough (we
don't skip draw calls while geometry shaders / tesselation is disabled),
but the point of this is just a quick hack to identify if these features
are in use.

Looks like we need to support compute shaders for some effects (specular
highlights, environment reflections... basically "physical lighting"
effects) in The Witcher 3. At least I hope this is the case and doesn't
turn out to be another wild goose chase...