Add option to clear render targets during frame analysis
If analysis_options=clear_rt is enabled render targets will be cleared
to black the first time they are set while analysing a frame. This makes
it much easier to see what objects are being drawn if the game has not
cleared the render targets ahead of time.

The downside is that a render target that is supposed to retain data
from a previous frame (e.g. sporadically updated global illumination or
reflections) will also be cleared , hence why this is an option.

TODO:
- Take unordered access views used by compute & pixel shaders into
  consideration (I suspect this is currently clearing the results of
  several compute shaders in the Witcher 3)
- Don't clear render targets if the game explicitly cleared them
- Clear the depth buffer
- Clear render targets already assigned at the start of the frame