Frame analysis: Consider UAVs as well as RTs when clearing RTs
Compute shaders use unordered access views, which might point to the
same Texture2D as a render target, so consider them as well when
deciding to clear a render target or not.

With this change the frame analysis of The Witcher 3 is correctly lit
again. Another clue that compute shaders are indeed responsible for part
of the lighting.