frame analysis: Dump Texture2Ds used as UAVs
This didn't give me any results in the Witcher 3, however I do now see
a gap in the draw count that wasn't present before at the exact point I
believe a compute shader would be running to calculate the lighting,
suggesting that there is indeed one there.

If it's not using a Texture2D resource it would make sense that nothing
was dumped, or I may just have a bug somewhere.