Fix logging macros to work properly in if/else statements without curly brackets
This fixes unordered access views not being dumped during frame
analysis, and 001062-UAV0-compute.jps seems to confirm my suspicion that
the lighting IS drawn in a compute shader. (plus another compute shader
looks awfully similar to the interior mask issue I could never track
down in FC4... could it be?)

Details
-------

This code was not being expanded as intended:

if (FAILED(hr))
	LogInfo("frame analysis: failed to create filename: 0x%x\n", hr);
else
	DumpResource(resource, filename);

It was expanded to:

if (FAILED(hr))
	if (LogFile)
		printf(LogFile, "frame analysis: failed to create filename: 0x%x\n", hr);
	else
		DumpResource(resource, filename);

This change makes it expand so the else statement is associated with the
correct if:

if (FAILED(hr)) {
	do {
		if (LogFile)
			printf(LogFile, "frame analysis: failed to create filename: 0x%x\n", hr);
	} while (0);
} else
	DumpResource(resource, filename);

(the do {} while (0) is a common construct used in these situations to
make sure the semicolon in the original code does not terminate the if
statement before the else)

n00b mistake on my part, d'oh!

Should also fix several logging messages not being displayed in error
paths.