Add a new marking mode of 'pink' as an option.
This sets the pixel shader output to always be hot pink (1,0,1,1), as a more visible and easier spot effect.  Sometimes going blank is not visible.

Tested and works in Alien and Witcher3.

Only drawback is it draws the entire bounding box for an effect, not just the pixels normally seen.

For the screenshots, this makes the documented shader much more obvious.