Ignore _reasm.txt files when looking for missing shaders to revert
Removing an assembly file and reloading was not working as expected. It
seems that the _reasm files were still being matched preventing the
revert from running on that file.

Blacklist _reasm files to prevent this from happening.

This may be better coupled more closely with the code that actually
reloads the shaders, but need to be careful we do the right thing in the
various code paths, so for now this is simpler.