Fix 32bit build due to missing calling convention in function pointers
The orignal functions used a typedef to specify __stdcall, but that
doesn't work in a function pointer since it ends up on the wrong side of
the brackets, so specify it manually. This isn't an issue on 64bit
builds because there is only one calling convention.

Probably explains why intellisense wasn't happy earlier.