Witcher 3: Fake hairworks depth output to fix transparency issue
This shader is copying the depth information from the right eye into the
left eye, causing further effects drawn (such as the sky) to draw over
the top of the hair making it look transparent.

The correct fix would be to work out why this is happening, but for now
faking the depth output seems to do the trick.