Frame analysis: First pass adding triggers to change options during run
This allows shaders and render targets (and soon UAVs and textures) to
specify analyse_options as well. When these are encountered during
analysis they will override the current analysis options, either for a
single draw call (default), or until the next trigger is hit (with
persist keyword).

This will allow the frame analysis to be used to:

- Analyse single draw calls that use a particular shader or render
  target without having to dump out render targets of unrelated draw
  calls

- Set triggers before and after a suspect part of the frame to enable
  dumping DDS files for only that part of the frame, taking
  significantly less time and disk space than if DDS files were enabled
  for the whole frame.