Fix bad buffer length calculation causing memory corruption in debug builds
Apparently I fail at basic arithmetic.

I had the filename and ext pointers the wrong way around in the buffer
size calculation, which lead to the string copy being told it had a
larger buffer than it really had.

This was not a problem in a production build as the buffers themselves
are the correct sizes and no overrun can ever occur.

In a debug build however, the string copy function poisons the buffer,
and since the length I gave it was wrong it corrupted unrelated memory.