Add ShaderOverride fake_o0 option to assign a dummy render target
This creates a render target that should be compatible with the assigned
depth target and assigns it as o0 in the output merger. This is useful
to influence driver heuristics and force a shader to run twice to fix
various one-eye issues.

This option can fix the MSAA hairworks in The Witcher 3, as the pixel
shader that is run to resolve the MSAA hairworks depth buffer into a
non-MSAA depth buffer is usually run with no render target assigned, and
as such the driver only runs it for the right eye creating a one-eye
depth buffer.