Fix incorrect format directive used during hunting next/prev causing garbage in logs
Size was originally a size_t to match that returned by set::size(). At
some point it was changed to an int, which no longer matched the %Iu in
the format string, causing the log file to read 64bits of a 32bit
variable and spew nonsense - we're lucky the stack pointer didn't get
corrupted and crash.

Change the format string to a %d to match the type change.