Skip decompiling unsupported shader types as a stop-gap measure to avoid crashes
At least one domain shader is causing a buffer overrun in the
decompiler, and it is likely that more bugs lie in wait.

For now prevent these new shader types from being decompiled when
export_hlsl is set so we don't unnecessarily reduce the stability of
3Dmigoto. Decompilation can still be attempted through marking.

As a side note, ReplaceShader() is far too complex. I barely glanced at
it before I spotted several bugs, but it's not worth fixing unless part
of an effort to refactor it into smaller more manageable functions.