Add other shader types to frame analysis filenames
This should give us a more complete picture of each draw call. Shaders
are only included in the filename when they are actually in use.

Changes the render target number to a format that resembles the register
name - o0 for a render target, u0 for a UAV and oD (short for oDepth)
for a depth target.