Open courierbold.spritefont from the exe directory, don't crash if missing
The Witcher 3 changes it's working directory after launch, causing us to
look for the overlay font in a different directory everything else, and
crash when we can't find it.

Use the same logic as the log files to always look for it in the exe
directory and handle the case where it does not exist by simply
disabling the overlay.