Witcher 3: Add toggle key for HBAO+
HBAO+ dies periodically in this game - the HBAO depth pass shader seems
to stop being called for the left eye. For now add a toggle key so it
can be disabled on the fly without having to open settings (which can be
inconvenient if it happens in a cutscene).

There's a few things to try when I get more time - faking a depth buffer
similar to how we faked a render target for the other one eye issue,
forcing the creation mode of the output render target to stereo, adding
more analysis check if the inputs to the shader are stereo.