First pass attempt at using the depth buffer as a shader resource
This copies the depth buffer that is currently set into a shader
resource. The idea is to allow UI elements to automatically adjust their
depth to fall on an object in the depth buffer, in the same way my auto
crosshair code works in DX9 games.

This will only copy the depth buffer in, not the stencil buffer - this
is due to the limitations of resource types that can be used as a shader
resource - none have both fields resolved simultaneously. If the stencil
buffer was ever needed, it would only be a very slight variation on this
code to select it instead of the depth channel.

At the moment this is limited to using the depth buffer that is
currently assigned to the requested shader, which won't work for all
games and requires a full resource copy for each UI element. Once this
approach is proved out we may improve that by allowing the depth buffer
to be fetched from elsewhere, or limited to being copied once per frame.

Later we can extend this to copy arbitrary resources between shaders.