Witcher 3: Experimental auto UI depth adjustment
This implements my auto crosshair code in HLSL to place UI elements at a
depth determined by iterating over the depth buffer. The code itself is
self contained in a function ready for copy & pasting. The near and far
clipping planes will need to be determined for each game (e.g. from a
constant buffer, or by guessing).

The X & Y coordinates can be determined either by the MVP matrix in the
UI shader (at coords 0,0,0,1), or just the output coordinates. The
former will keep all points on a single UI element at the same depth,
but sometimes is offset versus where we actually see the element (so
text will react to Geralt's depth when it is further to the right),
while the later can distort a given element, but will position it at
more or less the right depth.

~ is set to cycle between the three modes (fixed depth, use MVP, use
output coords).

Not all UI shaders are converted yet - this is enough for a proof of
concept, but I'm unsure if this is a good idea in this game or not.