Fix refcounting double release bug & set resource again
Adding caching at 5am leads to this kind of thing... The resource view
should stay alive while it is in the cache, otherwise we would double
free it after a resolution change.

It should also be set again whenever the shader is hit (though, we can
probably get away without that so long as the game doesn't ever use the
resource slot and we only inject into either vertex or pixel shaders...
I might rethink this).