frame analysis: Add filename_reg option to swap draw and register numbers
I've noticed that draw calls tend to use the same output registers for
the same purpose a lot of the time (e.g. o0 is usually the colour
output, o1 might usually be the normal buffer, etc).

In some cases, it might be desirable to have the filename such that it
will be sorted by register number instead of draw call to make it easier
to examine how an output changed throughout the frame.

For anything more sophisticated - use a bulk rename tool.