Fix compile issues to sync with new D3D_Shaders
- Set D3D_Shaders back to a static library

- Fix include directories for D3D_Shaders to pick up the files from
  HLSLDecompiler.

- Add a DISABLE_LOGGING define to the D3D_Shaders project which disables
  the Log* macros. D3D_Shaders now includes HLSLDecompiler, which
  expects our logging macros to work, but the logging infrastructure is
  not initialised in D3D_Shaders.

Note that since various configurations were removed from both the
D3D_Shaders and BinaryDecompiler projects, the active configurations in
StereovisionHacks.sln were updated. It's not clear what is best to do
here, if we need to restore the old active configurations, the easiest
way is:

git checkout xxxxxx^ -- StereovisionHacks.sln

Where xxxxxx is the sha of this commit.