Expand the number of ini params to 8x4
Expanding the number of ini parameters allows us to have more tunables
in our fixes, and opens up the possibility of passing in filtering
information from 3Dmigoto in these parameters in the future.

The first lot of ini params are set via x=, y=, etc. as before

Additional ini params can be set via x1=, y1=, x2=, etc.

Currently this is limited to 8x4, but it is now very easy to adjust via
the INI_PARAMS_SIZE constant in globals.h