Delete all project specific build targets.
These are just cruft, and wouldn't hurt anything, but they show up in the UI are not necessary.

These are hand edited in the XML files, because VS is too stupid to have a way to delete these from the UI.

Ignoring project files that are not currently actively used, like d3d_compiler.