Create built target "Zip Release" for DirectXTK project.
This came in after Zip Release, and thus needed to be manually added to the project.

This needed to be straightened out because we were building the Release version for only this one project, which broke the goal of having an independent build machine target that is not prone to modification.