Clean up cruft in the D3D_Shaders subproject.
This has been built in a funny way several times, and this unifies all the settings to a project global technique.

Cleaned out all unused includes, all unused libraries.

Cleaned out all unused PreProcessorDefinitions.

This is the simplest case now for building this project, with no accidental includes of 8.1 headers.