Clean out the cruft in the DX11 project.
All includes are now as simple as possible, including only the bare minimum we need to compile and build.

Also avoids the accidental use of Win8.1 header files by not including system wide headers.

Also removed the use of ComPtr's in Hunting, for the ScreenGrab function.  Not worth the extra includes.

Compiles and builds.