Add a DLLMain for d3d_compiler46 project.
This will get loaded at static load time, which allows us to load our d3d11.dll and thus get our hooks installed earlier than usual.

This seems to be necessary for Batman AK, as it has the d3d11 set to delay load, which meant that dxgi.dll was loaded before we got a chance to hook the factory, and thus missed the CreateSwapChain call.

This now works to enable shader hunting.  Overlay is missing in action, but that can easily be a graphic problem.