Frame Analysis: Add dump_cb option to dump constant buffers
This can be used to identify what constant buffers are bound to each
type of shader in the pipeline, and examine their contents.

Filename is changed slightly for textures and constant buffers - rather
than implying which shader it was bound to to by being the only shader
in the filename, it is now explicitly listed before the register and all
bound shaders are listed. That should make the files sort more logically
and provide more information without having to dump the render targets
as well.