Add InitDrawState to Overlay.
Project Cars and Batman Arkham Knight were not showing the overlay, because the drawing environment the game left at Present() did not work.

Rather than rely upon games to have a usable state, this now sets the state deliberately, for the parts we know we need for DirectXTK.
This includes setting the backbuffer as the rendertarget, and a wide open viewport.

Those pieces are also saved and restored over the DrawOverlay call.

Removed the check for viewports being 0, as we now specifically set the viewport to avoid that as well as any clipping from the game.