Add text parsing for resinfo_indexable.
The Batman and Witcher games both stripped headers of even reflection info, which makes the James-Jones parser return null resource information.

This handles that scenario by doing a text parse of the input line, and generating the corresponding instruction.
Only works for a limited subset of resinfo, but handles the ones we know about.