Add expand_region_copy option to TextureOverrides
This option is used to expand the region passed to
CopySubresourceRegion() to cover the whole width of the texture. This is
used in Lichdom Battlemage to resolve issues where transparent effects
receive a partial copy of the back buffer, but it's limits don't cover
the entire effect in stereo (much like scissor clipping).