Restore resolution match test for texture hashes & add multiples of resolution
This restores the code that special cases textures that match the back
buffer resolution so that the hashes will be consistent regardless of
the resolution being played at.

It also adds special cases for *2, *4, *8 and /2 multiples of the
resolution. *2 is needed for Far Cry 4 and others should help in various
cases.