Add SkipPacking routine to avoid fxc packing errors.
fxc will pack two TEXCOORD that use e.g. xy and zw in different registers, back into a single register.  This overly aggressive packing changes the input or output signature of a shader and thus breaks the shader.

This SkipPacking routine will try to determine if upcoming textures are used for a legitimate packing, using the same register.  If not, then it will force the format to float4 for any that might be overly packed.

This forces fxc to back off, and allow the signature to be unchanged. The side-effect is that we get warnings for the extra parameters that are not assigned.

After numerous experiments, this forcing to a float4 is the only technique that worked.

This changes both ParseInputSignature and ParseOutputSignature to force any TEXCOORD to not be incorrectly packed.

Tested in Batman, and it fixes all the problems seen there, without affecting legitimate packing.