Add ability to get the resolution to match texture hashes from the depth buffer
Lichdom creates a swap chain the same size as the screen's native
resolution regardless of what resolution is set in game, then upscales
to it.

This means we can't use the swap chain dimensions to special case
texture hashes that match the resolution, so we have to find an
alternate and reliable means. It stands to reason that a rectangular
depth buffer is likely to match the resolution set in game, so add it as
an alternate means to match the resolution.