Match depth/stencil textures when resource is created instead of view
Waiting until the view is created means we won't special case the
hash for the depth buffer itself. There probably won't be too many times
were we need to do that, but we can solve the issue by checking the bind
flags when the texture is created rather than waiting for a view to be
created.