Work around crashing bug in UPlay and Origin games
Both these systems create a DX10 device (Origin for the overlay, Uplay
for not sure what). Turns out that DX10 just creates a DX11 device with
feature level 10.0 these days, but that crashes for some reason.

For now, fail the creation of any feature level 10.0 devices to avoid
this crash.