Frame analysis: Fix any dump_*_dds affecting other types
If any dump_*_dds (or dump_*_jps) was enabled it would enable dumping
that type of file for other dump options as well (e.g. if dump_tex_dds
and dump_rt_jps was enabled both would end up dumping dds and jps
files). Pass the mask for the type of object being dumped up the call
stack so Dump2DResource can determine the correct action based on the
type and options.