Add deny_cpu_read texture override option to prevent CryEngine's culling
This option will replace the buffer of a mapped resource that the CPU is
reading with an empty buffer, thereby preventing the game from reading
the resource.

It is intended for CryEngine games that read (depth?) buffers back from
the GPU to determine if an object was drawn or not to decide to cull it
from future frames, which does not work in stereo making objects
disappear and flicker.