Implement texture filtering
This implements a mechanism that can be used for texture filtering. It
works be passing a value in through the ini params that maps to a
defined texture.

In order to use this, a ShaderOverride feature sets one of the ini
parameters to the shader type and texture slot to be checked. For
example, to check texture slot 0 of the pixel shader, you would do
something like:

[ShaderOverrideHUDAndCrosshair]
Hash = ...
x1 = ps-t0

The value passed into the parameter is obtained from the
[TextureOverride] sections. If no section exists for the texture, the
parameter will be set to 0. If a section does exist, the parameter will
be set to 1 (by default):

[TextureOverrideCrosshair]
Hash = ...

If more advanced filtering is required, a custom index can be set on a
per-texture basis, for example:

[TextureOverrideCrosshair]
Hash = ...
filter_index = 2
[TextureOverrideDamageIndicator1]
Hash = ...
filter_index = 3
[TextureOverrideDamageIndicator2]
Hash = ...
filter_index = 3
[TextureOverrideMap]
Hash = ...
filter_index = 4