Fall back to using full paths if LoadLibrary("original_xxx.dll") fails
Origin's IGO32.dll inserts a hook in ntdll.dll's LdrLoadDll, which can
interfere with our hook in LoadLibraryExW(). In particular, depending on
which library got loaded first it may see our faked filename and attempt
to GetModuleHandle/GetModuleHandleEx it (which fails as we loaded the
module with the real name, which is different to the one it saw), and
even if we hook those to return the redirected module it them attempts
to load the fake filename directly, which fails and ends up with the
entire LoadLibraryEx call failing.

While annoying, this is most likely not malice, rather it is just a
pitfall of hooking and a code path they did not expect or test.

For now, if the LoadLibraryEx call fails, try again with the real full
path since IGO will let that through. We have to suppress the redirect
for that call to make sure we don't get a reference to ourselves as that
would quickly lead to an infinite recursion, crashing once the stack
runs out.