Change buffer parsing to handle long names.
The buffer var names can be 'too long', and if so, the disassembler will smash the comments into the ; at the end of the line, which made our name parsing fail, as %s %s would not find white space.

This forces it to %s %[^;] to find anything except a semicolon as the terminator.

Not tested, but will do that in the next check-in.