Fix bug introduced for dcl_constantbuffer in named headers case.
When adding support for the no-header case of Witcher3 and Batman, we add a series of names like cb2, cb3 for constant buffers where we are unable to fetch the names.

This check was wrong, as it would also generate these blank named cb* buffers if the offsets into the buffer were non-zero.  Turns out you can have starting offsets that are non-zero, so the check for the 0th entry for a mCBufferData was wrong.

Rather than look at that array of names, this is switched to mCBufferNames.empty() as the check, because in the no-header case, all the cbuffer declarations are stripped, and there will be no names.

This fixes a problem seen in Crysis3, where we were not finding the named structures properly, and defaulting to the generic cb* form.