Rewrite SetShader for performance.
Rewrote the routine, but hopefully did not change behavior.  It's got a lot of complicated features though.

This removes all but one of the EnterCriticalSection, and puts that where a map is actually modified, and nothing else.

Reorders the subpieces so that we can just modify the shader passed in, and set that shader in the same spot in the routine for all of them.

Does an optimization to see if the mShaderOverrideMap is empty or not, and if it's empty, skip doing lookups that take measurable CPU time when profiled.

Added several other .empty() checks for other hot spots in the code, so that unless we are using one of the esoteric features of 3Dmigoto, we don't pay any CPU price. Specifically TextureOverrides, ShaderOverrides, or BufferOverrides.

After this change, the profile demonstrates we recovered all that CPU time.  1.5% CPU to 0.8% CPU now.

I do not believe the remaining items can be trimmed further.