Make the Overlay Init use original DX11 objects.
Was getting a crash from OMSetRenderTargets, and it seemed like a bad idea to call through our code to setup the DirectXTK drawing environment.

This does GetOrig on each of the pieces to make sure we are just setting up the environment without capturing those calls or logging them.

Lower GetOrigSwapChain to LogDebug as it's now called for every frame.