Add the crc32c checksum code.
This is based on Mark Adler's version, and altered to compile in VS by Robert Važan.

Code is unchanged, other than to remove the precompiled dlls and .net support.  We want to compile the code as part of our project, not link to a library.

The reason to use this checksum is because it is hardware accelerated using SSE 4.2 instructions.