Initial changes from fnv_64_buf to crc32c_hw for texture hashes.
The motivation for changing the hash function is to use one that is much faster than the fnv_64_buf.  This crc32c_hw uses a CPU instruction in SSE 4.2 that does the crc32c function, and thus runs 30x faster than fnv.

If textureoverrides are in use, this will provide a measurable bump in performance, as we then need to hash every texture seen.

This is the basic structure of replacing the hash calculation with the new variant.

Adds the crc32c_hw function wrapper for the unchanged crc32c_append. The main reason to do this is to avoid having to cast the usual void* windows pointers to the uint8_t* that the crc library inexplicably wants.  Adler's original code used void*.

This keeps the try/catch block here too, so that if the crc32c_append hits some missing data, we'll catch it and return zero in all cases.