Initial setup for crc32c, including validation code.
Set up to calculate twice, once with crc32c, once with fnv_64, so we can compare collision rates.

In GTA5 and Witcher3, the crc32c has a slightly lower collision rate.

This variant is set to use magic constants for the textures that are screen resolution sized, to avoid the dynamic hash numbers.