Clean up CheckForSpecialCaseTextureResolution.
Name here is hard, but renamed to be slightly more clear AdjustForConstResolution.

This allows us to clean up the CreateTexture*D functions to not have to check resolution, and do it only in the CalcTexture*DDescHash function instead.

Would really rather have these utility functions out of the HackerDevice object, but Globals+Utils is a bit of a mess and makes that overly hard.  (Read as: too many dependencies and circular references, so, later.)