Remove collision checking code for Texture2D hash.
No real point in even having this in a Debug build, because there's too many 'normal' collisions to watch out for, and thus noisy data.

Given that we think the crc32c is working adequately, we don't need this.  Always here in the project if we need to check it again sometime.