Remove early exit from CreateTexture2D to fix get_resolution_from=depth_stencil
With the early exit placed where it was, the depth/stencil resolution
check could get skipped if the game did not define any TextureOverrides
and had hunting disabled. Currently this feature is only used in games
that define TextureOverrides, but it is a potential gotcha for someone
down the road. Remove the early exit to resolve the issue.

See also the commentary here:
https://github.com/bo3b/3Dmigoto/commit/267b225e7c3b2df42ec7b53b85f476622ea806ee#commitcomment-12848592