Wrap all of Buffer, Texture*D calls with HackerBuffer.
This is experimental, but wraps all ID3D11Resource creation with appropriate C++ objects to make them changeable later.

Right now this crashes, because we are likely passing in a HackerTexture2D to SetShaderResources, instead of an ID3D11Resource, and we need to pass the original there and all other uses of the Resources.