IniHandler: Preserve order of settings in a section
The order operations are performed in a shader override will become
important for an upcoming feature, and how it interacts with other
features.

For example, one of the planned features will add the ability to run a
custom shader on the GPU to enable advanced decisions on a game by game
basis that I don't want to hard code in 3DMigoto itself. In particular,
I intend to use this to copy two sets of matrices around in DreadOut for
the two distinct render targets that game uses. For that to work, it is
important that an ini parameter has been loaded with the render target
dimensions prior to running the custom shader that copies the matrices.

I could simply decide on a fixed order to process each feature to make
sure that the ini parameter is assigned first, but it's not clear that
will always be the correct order, so it seems better to allow the
shaderhacker to define the order in the d3dx.ini.

Changing this now while the order doesn't matter for any existing
feature means we won't break anyone later when the order does matter.