Introduce Shader override command list
This adds a vector of commands to be carried out to each ShaderOverride.
This will allow us to process actions from a ShaderOverride section in
order. This is in preparation for some upcoming features where the order
will be important.

It also means we can potentially check less things each time a
ShaderOverride is triggered in a frame, though that is not likely to be
very significant and is not the main goal.

The new parsing code also has the added benefit that we can now warn
about unrecognised lines in a [ShaderOverride] section since we are now
checking every line.