Migrate ShaderOverride IniParams overrides to command list
This proves out the command list approach, and is in preparation for
being able to inject new draw calls in the future as the ordering of
this feature will become significant at that point.

There should not be any noticeable change in behaviour of existing
features.

Most of the other features do not affect state in a way that would make
the order important, or are planned to be superseded by an upcoming
feature.