First pass adding ability to copy constant buffers between shaders
This is a first pass to implanting arbitrary resource copying. Currently
it is limited to copying constant buffers by reference between shaders
of the same draw call, but that is enough to give a vertex shader access
to a constant buffer of the corresponding pixel shader, which can be
used to solve the bloom depth in Mad Max and perhaps other games.

The code is packed full of TODO comments to support other resource
types, perform full copies, and copy resources between unrelated
shaders.