Hook up texture copying between shaders by reference
This is still somewhat limited as it can only grant a shader access to a
texture that another active shader has access to in the same draw call,
although the texture will remain bound so if the game leaves it alone it
may be accessible from later draw calls.

This was low hanging fruit and gets some of the structure ready for
dealing with views.