Hook up render + depth targets and implement support full copying
This expands the functionality somewhat. Full copying is now supported,
which is necessary for these types as they are bound to an output and
therefore cannot be bound as an input so copying by reference won't
work.

Currently a new resource will be created each time and the old one
discarded. We will want to keep these around so long as their
descriptions don't change for better memory management on the GPU and
possibly increased performance.

This should supersede the old depth_input feature, e.g. with:
vs-t110 = oD

Note that *setting* a render / depth target has not been tested yet as
that functionality won't be very useful until we can create entirely new
resources.