Resource Copying: Use EnsureNotTypeless() instead of source view format
I was getting an invalid argument error when trying to copy the
depth/stencil buffer into a shader resource. This is most likely due to
the fact that the formats that support a depth and a stencil channel
simultaneously can only be used for a depth/stencil view. For now switch
to using EnsureNotTypeless which already handles that case.

Later I might consider applying the format when the new view is created
instead.