Drop superseded depth_input feature
This feature has only ever been used experimentally in Witcher 3, and is
not currently used in any complete fix. It was reported to have a memory
leak, and is not very flexible.

The new arbitrary resource copying feature can already be used to do the
same thing, but will become much more flexible in the future (e.g. it
will be able to copy resources from unrelated shaders). Given that,
scrap this feature.