Resource copy: Hook up vertex buffers, index buffers and stream output
This allows vertex buffers, index buffers and stream output buffers to
be copied around as desired.

As I see it, the most interesting possibility this opens up is the
ability to pass a vertex buffer into a shader as a constant buffer. For
example, a vertex shader could make use of this to look up other
vertices in the same draw call. This could allow us to find the bounding
box or center of a UI element, which could be useful if we wanted to
automatically adjust the depth of a UI element based on the depth
buffer, so that all vertices could find the same point (or points) on
the depth buffer to sample while avoiding the offset problem we ran into
in the Witcher 3 when we only used the MVP matrices to get the position.

We might want to follow this up and allow additional information about
the vertex buffer to be passed into a shader. e.g. we might want to get
the strides + offsets values in, which we could store in the IniParams
easily enough. However, we might also want to pass information about the
ID3D11InputLayout (e.g. so we can work out where the POSITION0 values
start as opposed to TEXCOORD0, etc), but there is currently no way for
us to query that at run time as far as I can tell. In the future we
could potentially wrap CreateInputLayout() to get this info.